recipe_service/requirements.txt
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
Flask==2.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.26.0
urllib3==1.26.20
Werkzeug==2.0.1
